hall nawh

+++ Szabdan/Asztali/Öröklödes feladat/2024.11.19. Szuperhos/2024.11.19. Szuperhos/Szuperhos.cs
@@ -9,6 +9,6 @@ namespace _2024._11._19.Szuperhos
     internal interface Szuperhos
     {
         bool legyoziE(Szuperhos szuperhos);
-        int mekkoraAzEreje();
+        float mekkoraAzEreje();
     }
 }
